refactor: GET 상품 리뷰 count 제거(Product 테이블에 추가됨)
