+++ src/product/review/product-review.service.ts
@@ -160,23 +160,4 @@ export class ProductReviewService {
     };
     return reviewsData;
   }
-
-  async count(productId: number) {
-    const product = await this.prismaService.product.findUnique({
-      where: {
-        id: productId
-      }
-    });
-
-    if (!product) {
-      throw new NotFoundException('Product not found');
-    }
-
-    const count = await this.prismaService.review.count({
-      where: {
-        productId
-      }
-    });
-    return { count };
-  }
 }
