+++ src/review/review.controller.ts
@@ -16,7 +16,7 @@ import { plainToInstance } from 'class-transformer';
 
 import { AccessTokenGuard } from '@/auth/guard/access-token.guard';
 import { FindReviewsDto } from '@/product/review/dto/find-reviews.dto';
-import { ReviewCountDto } from '@/product/review/dto/review-count.dto';
+import { ReviewCountDto } from '@/review/dto/review-count.dto';
 
 import { ReviewDto } from './dto/review.dto';
 import { ReviewsDto } from './dto/reviews.dto';
