+++ test/controller/review.integration-spec.ts
@@ -4,7 +4,7 @@ import * as request from 'supertest';
 
 import { AuthModule } from '@/auth/auth.module';
 import { PrismaService } from '@/prisma/prisma.service';
-import { ReviewCountDto } from '@/product/review/dto/review-count.dto';
+import { ReviewCountDto } from '@/review/dto/review-count.dto';
 import { ReviewDto } from '@/review/dto/review.dto';
 import { ReviewsDto } from '@/review/dto/reviews.dto';
 import { UpdateReviewDto } from '@/review/dto/update-review.dto';
